Move NotificationHandler to common
To be able to put this as dependency in AssignmentActionManager which is
now responsible for cancelling notification. That way, when a task is
snoozed or completed from within the app, the notification can be
cancelled.
Also added new methods to get new reminder time on snooze
Also fixed a bug where edits would be allowed for others' assignments
And finally, snoozing by tomorrow would result in reminder being shown
at 8AM the next day