Attempt to fix assignments getting overwritten (#27)
When an assignment is marked as done from notification, it sometimes
results in waking up the app and triggering the "upload existing,
download new assignments" worker. Since db access is not serialized,
worker sometimes picks up the old list of existing assignments. This
results in the assignment never getting uploaded and then also it
getting replaced by what's downloaded from the backend.

Changed the logic to not replace existing row with what we get from
backend since that would likely never be intended (at least right now).
If a row exists, then it is skipped and not saved again to db.