Delete all todo assignments when fetching new
There was an issue in the previous logic where only the ones that get
uploaded were cleared off from database. This means that only the ones
for the user logged in, would ever get deleted from the database, as
the user cannot mark other users' assignments as completed and hence
they would never get marked as ready for upload.

This changes that and deletes all assignments that are in TODO state and
replaced by the ones sent by backend.