Fix duplicate notification
There was an issue where I was adding an alarm to the db but it was getting added with a new notification id every time the alarm got rescheduled.

Which meant that a duplicate notification was being shown.

Fix was to use existing row's notification id if it existed otherwise let SQLite generate a new one automatically.