Change release version format
Automating release version change so changing the format.
Going forward, version name be something like 6.6 and version code
will be version name times 10. This makes it easier to generate the new
version name and code by incrementing 0.1 to name and can be automated.
Next version for the above example will be 6.7 - 67.