Fix rescheduling bug from time zone change
When added "sync on time zone change" functionality, didn't set up
alarms for future assignments as they were getting their notification id
as -1. Which in database layer meant that it would be inserted as a new
row.

However, if a row for that assignment already existed it won't be
inserted at all plus any other row changes that were going in with the
future assignment, would  also be scrapped as the whole change was being
done in a transaction.

So changes from sync where all reminder times should've been reset,
wasn't taking effect at all.

Fix was to use existing notification id for future assignments if it
existed and only fallback to -1 if it didn't